espace user pour les produits

+++ src/Controller/ecommerce/ProduitController.php
@@ -7,6 +7,7 @@ use App\Entity\Caracteristiques;
 use App\Entity\CategorieProd;
 use App\Entity\Date;
 use App\Entity\Produit;
+use App\Entity\User;
 use App\Form\ecommerce\ProduitType;
 use App\Repository\CategorieProdRepository;
 use App\Repository\ProduitRepository;
@@ -248,12 +249,19 @@ class ProduitController extends AbstractController
     public function delete(Request $request,EntityManagerInterface $manager,FileUploader $uploader)
     {
         $id = $request->get("idProduit");
+
         /** @var Produit $produit */
         $produit = $manager->getRepository(Produit::class)->findOneBy(["id"=>$id]);
         if(!$produit)
             die();
 
-        foreach ($produit->getImages() as $imgeName)
+        /** @var User $user */
+        $user = $this->getUser();
+        if(!$user->isAdmin())
+            if($produit->getClient() !== $user)
+                die("vous ne pouvez pas acceder à cette page");
+
+        /*foreach ($produit->getImages() as $imgeName)
             $uploader->deleteFile($imgeName,"produit");
 
         $manager->remove($produit);
@@ -263,7 +271,7 @@ class ProduitController extends AbstractController
         foreach ($produit->getAvis() as $avis)
             $manager->remove($avis);
 
-        $manager->flush();
+        $manager->flush();*/
 
         return $this->json(["success"=>["produit supprimé"]]);
     }
